% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{biomarker}
\alias{biomarker}
\title{Calculate Biomarkers}
\usage{
biomarker(
  m_ind_traits,
  a_coef_ukpds_ind_traits,
  a_coef_ukpds_other_ind_traits,
  biomarker_eq,
  time_step
)
}
\arguments{
\item{m_ind_traits}{A matrix containing patient characteristics over time.}

\item{a_coef_ukpds_ind_traits}{A 3D array of coefficients used for calculating risk.}

\item{a_coef_ukpds_other_ind_traits}{A 3D array of coefficients \verb{c("lambda", "rho", "death"} used for calculating risk.}

\item{biomarker_eq}{A character string specifying the health outcome equation (e.g., "ihd").}

\item{time_step}{An integer indicating the row in \code{m_ind_traits} to use for calculations.}
}
\value{
The updated biomarker is stored.
}
\description{
This function calculates patient-specific factors to predict the time path of
a biomarker (linear progression of risk factors).
}
