% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ukpds_bootstraps}
\alias{df_ukpds_bootstraps}
\title{UKPDS Model Bootstrap Replicates for PSA}
\format{
A tibble (data frame) with 4001 rows and 48 columns:
\describe{
\item{Rows}{Each row is a single bootstrap replicate.}
\item{Columns}{Each column represents a specific model parameter, such as
\code{re_hb_fem} (the female coefficient for the HbA1c risk equation) or
\code{re_sbp_cons} (the intercept for the SBP risk equation).}
}
}
\source{
Generated from a bootstrapping procedure of the original UKPDS Outcomes
Model 2 risk equations.
}
\usage{
df_ukpds_bootstraps
}
\description{
A data frame containing 4,001 sets of bootstrapped coefficients for the UKPDS
risk equations. Each row represents a complete parameter set for one iteration
of a Probabilistic Sensitivity Analysis (PSA).
}
\keyword{datasets}
