% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{initialize_patients}
\alias{initialize_patients}
\title{Initialize baseline values for multiple patients}
\usage{
initialize_patients(num_patients, m_ukpds_pop, m_ind_traits)
}
\arguments{
\item{num_patients}{Numeric. The total number of patients to process.}

\item{m_ukpds_pop}{A Matrix containing patient characteristics.}

\item{m_ind_traits}{A matrix to store patient data.}
}
\value{
The updated matrix with initialized patient data.
}
\description{
Initialize baseline values for multiple patients
}
