% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{mortality}
\alias{mortality}
\title{Calculate Mortality Events for a Given Time Step}
\usage{
mortality(
  m_ind_traits,
  m_other_ind_traits,
  a_coef_ukpds_ind_traits,
  a_coef_ukpds_other_ind_traits,
  time_step
)
}
\arguments{
\item{m_ind_traits}{A matrix containing patient-level data, including health event history.}

\item{m_other_ind_traits}{A matrix containing lmabda, rhos and death.}

\item{a_coef_ukpds_ind_traits}{A coefficient matrix used in Gompertz and logistic event calculations.}

\item{a_coef_ukpds_other_ind_traits}{A 3D array of coefficients \verb{c("lambda", "rho", "death"} used for calculating risk.}

\item{time_step}{An integer specifying the time step at which mortality should be calculated.}
}
\value{
The updated \code{m_ind_traits} matrix with the mortality status recorded for the specified time step.
}
\description{
This function calculates mortality events for a given time step
based on new health events and medical history using Gompertz and logistic models.
}
\examples{
\dontrun{
m_ind_traits <- mortality(m_ind_traits, a_coef_ukpds_ind_traits, time_step = 5)
}

}
