% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsimV.R
\name{mortalityV}
\alias{mortalityV}
\title{Calculate Mortality Events for a Given Time Step (vectorised)}
\usage{
mortalityV(
  m_ind_traits,
  m_other_ind_traits,
  m_other_ind_traits_previous,
  m_coef_ukpds_ind_traits,
  m_coef_ukpds_other_ind_traits
)
}
\arguments{
\item{m_ind_traits}{A matrix containing patient-level data, including health event history.}

\item{m_other_ind_traits}{A matrix containing lmabda, rhos and death.}

\item{m_other_ind_traits_previous}{A matrix containing preceding cycle lmabda, rhos and death.}

\item{m_coef_ukpds_ind_traits}{A matrix of coefficients for individual‐level risk equations.}

\item{m_coef_ukpds_other_ind_traits}{A matrix of coefficients \verb{c("lambda", "rho", "death"} used for calculating risk.}
}
\value{
The updated \code{m_ind_traits} matrix with the mortality status recorded for the specified time step.
}
\description{
This function calculates mortality events for a given time step
based on new health events and medical history using Gompertz and logistic models.
}
