% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{update_all_biomarkers}
\alias{update_all_biomarkers}
\title{Update Multiple Biomarkers in a Transition Matrix}
\usage{
update_all_biomarkers(
  m_ind_traits,
  a_coef_ukpds_ind_traits,
  a_coef_ukpds_other_ind_traits,
  time_step,
  next_row
)
}
\arguments{
\item{m_ind_traits}{The patient trace, a matrix containing patient data with biomarker and event columns.}

\item{a_coef_ukpds_ind_traits}{A 3D array of coefficients used for calculating risk.}

\item{a_coef_ukpds_other_ind_traits}{A 3D array of coefficients \verb{c("lambda", "rho", "death"} used for calculating risk.}

\item{time_step}{An integer representing the current time step.}

\item{next_row}{An integer indicating the row in \code{m_ind_traits} to update with new biomarker values.}
}
\value{
The updated transition matrix \code{m_ind_traits} with new biomarker values in the specified row.
}
\description{
This function updates multiple biomarker values in the transition matrix for a given time step.
}
\examples{
\dontrun{
# Example usage
m_ind_traits <- update_all_biomarkers(
  m_ind_traits,
  a_coef_ukpds_ind_traits,
  time_step = 1,
  next_row = 2
)
}

}
