% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{update_coef_with_bootstrap}
\alias{update_coef_with_bootstrap}
\title{Update a Coefficient Data Frame with a Bootstrap Replicate}
\usage{
update_coef_with_bootstrap(
  df_ukpds_coef = microsim.tutorial::df_ukpds_coef,
  df_ukpds_bootstraps = microsim.tutorial::df_ukpds_bootstraps,
  i
)
}
\arguments{
\item{df_ukpds_coef}{A data frame containing the model's coefficients. It must
have a character column named \code{Parameter} and columns for each biomarker
(e.g., "hba1c", "sbp").}

\item{df_ukpds_bootstraps}{A data frame where each row is a bootstrap replicate and
columns correspond to the bootstrapped parameter estimates (e.g.,
"re_hb_cons", "re_sbp_fem").}

\item{i}{An integer specifying the row index of the bootstrap replicate to use
from \code{df_ukpds_bootstraps}.}
}
\value{
A modified \code{df_ukpds_coef} data frame with updated values from the specified
bootstrap replicate.
}
\description{
This function takes a main coefficient data frame and updates its values for
a set of biomarkers using a single row (a specific bootstrap replicate) from
a bootstrap results data frame.
}
\examples{
\dontrun{
# Assume df_coef is your main coefficient data frame and df_boots contains
# 100 bootstrap replicates. To update the coefficients using the 5th
# bootstrap replicate:

updated_df_coef <- update_coef_with_bootstrap(df_coef, df_boots, 5)
}
}
