% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsimC_.R
\name{update_health_eventsC_}
\alias{update_health_eventsC_}
\title{Update Health Events Over Time Steps (vectorised calling C++ functions)}
\usage{
update_health_eventsC_(
  m_ind_traits,
  m_coef_ukpds_ind_traits,
  m_coef_ukpds_other_ind_traits
)
}
\arguments{
\item{m_ind_traits}{A matrix to store patient data.}

\item{m_coef_ukpds_ind_traits}{A matrix of coefficients for individual‐level risk equations.}

\item{m_coef_ukpds_other_ind_traits}{A matrix of coefficients \verb{c("lambda", "rho", "death"} used for calculating risk.}
}
\value{
The updated transition matrix \code{m_ind_traits} with new biomarker values in the specified row.
}
\description{
Update Health Events Over Time Steps (vectorised calling C++ functions)
}
